requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
Django==3.1.2
django-configurations==2.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-qiniu-storage==2.3.1
django-redis==4.12.1
djangorestframework==3.12.1
idna==2.10
mysqlclient==2.0.1
Pillow==8.0.1
pycrypto==2.6.1
PyJWT==1.7.1
pytz==2020.1
qiniu==7.3.0
qrcode==6.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
